Bug fix not required degree courses need conditon
During the EoC rehearsal it was found that candidates with a uk degree
caused an error when choosing degrees with `not_required` degree grades.
This condition was missing from the hash and was not caught by the guard
clause. This PR adds the condition so that the correct text is
displayed.