Update equality and diversity error messages
This brings the error messages in line with the GOV.UK Design System guidance, which uses 'Select' for radios instead of 'Choose'. This is particularly important here as 'Choose' is a loaded word to use in the context of disability, ethnicity and sex/gender.

Adding " or ‘Prefer not to say’" alerts users to this option if they select a radio option because they didn't want to answer the question.