Add training_provider filter to the application filter
At the moment i've had to implement a workaround to get it to
only show on the provider UI as the two interfaces are tightly
coupled.

I'm going to implement the accredited_provider filter and
then i'll refactor the shared logic out into a module and split
this into two separate filters.