Don't use the app form to determine which cycle to carry over to
At the moment, we use the current applications recruitment_cycle_year
to determine which cycle to carry over the application to.

It does this by taking the current applications recruitment_cycle_year
and adding 1 to it.

This works fine if someone is carrying an application over one cycle.
However, if the application form is two years old, this causes two
carry overs to happen.

Instead we should just figure out which cycle they should be in by
using the cycle timetable to establish where they are in the cycle.

If it's after the apply1 deadline, but before cycle launch, it should
be carried over the next cycle. If not then just carry over to the
current cycle.