Ensure candidate_api_updated_at is touched
When a candidate adds a course choice, reference or qualification it
triggers a touch on the application form. This causes the transition of
state between `unsubmitted_not_started_form` and `unsubmitted_in_progress`
to be missed unless we catch the update when the object
is created.

As the touch doesn't trigger the around_save method on the application
form.