Add splat args for other data_for_export methods
We call the various export classes dynamically from the DataExporter worker. In some cases we need to pass additional arguments to the underlying
export class, eg. when we want to run an export in the context of the current cycle.
This commit adds a splat argument to the classes which do not need to use the passed export_options.