Update recruitment cycle year after create
There appears to be a race condition in the ApplicationForm Factory which happens adding application choices using :application_choices_count and setting :recruitment_cycle_year to the previous year.
The :completed_application_form trait defaults to the current
recruitment cycle and overriding this to a previous year can
sometimes cause the issue when called with :application_choices_count.

It's assumed that this is a callback race condition between the factory
and the touch callback as the failure is not consistent.

It's easiest to replicate with the default :references_count of 2.

This commit wraps the updates to the form which add application choices
and references in a with_unsafe_application_choice_touches block.