Use Faker::Alphanumeric.unique for provider code
It's possible that Faker::Alphanumeric.alphanumeric will return the same code twice, this will cause the Provider factory to return an existing instance instead of creating one.
This allows for the side-effect that FactoryBot.create(:provider) will not always create a new record and we are seeing test failures where FK constraints on org permissions are
being hit.