Freeze start time of tests
There's an inconsistent test around the number of jobs run which seems to vary depending on the time the test is run.
This commit freezes the test run at the start of the current hour as the job count is based on hourly jobs at 5-15 mins past the hour.
The tick frequency of the test suite is also increased so that we always
call the clock schedule twice a minute.