Fix insufficient privilege errors during deploys
These occur because database credentials are rotated with every deploy
as part of dynamic service key binding. This commit creates a persistent
'postgres' service key, the credentials of which are exposed to Rails
instances via DATABASE_URL env var.

This commit also removes the dynamic re-binding of redis instances,
which is not needed as we have already switched to our self-managed
service keys for both Redis instances.