Add the ExternalReportApplicationsExport
This does quite a lot of things and is fairly complex.

It builds a base hash with all the possible permutations:
 - For primary courses thats all the primary subjects
 - Secondary has all the secondary subjects
 - Further educations subject can only be further education afaik

Testing this is fairly difficult due to the sheer amount of permutations.
Using an array of hashes with dummy data like we normally would, would
make the test file absolutely massive. To try and get round this i've
done the following:

1. Create an array of hashes in the desired format, but with 1
difference. I've added a key, based on the rows values.
2. Create courses and populate the corresponding row by accessing it
with the unique key.
3. Remove the key so the array of hashes is in the desired state

There may well be a better way to do this and i'm very open to ideas.