Geocode address after update
When a candidate carries over their application, multiple requests to
geocode an address are fired because the carry over process involves a
cascade of record 'touches'. When we check if the address has changed,
this resolves to true every time as the record has at that point not
been reloaded with the updated address details. This then means that
sidekiq is swamped with duplicate requests to geocode the same address
which can then cause the Google Geocode API to return an 'OverQueryLimit' error depending how much is is
being hammered at the time.

Moving the geocode process to the 'after update' callback, fixes this
issue.