Region code backfill should process 50 applications at a time
There was a logic error in the original code that would have caused each
application to be processed as a single batch with a 5 second gap
between each application. This would take more than 24 hours to backfill
the production DB rather than spreading it out over about an hour as
intended.