Ensure that the MonthlyStatistics do not show until the publish date
TAD need time to QA our data before we publish it. Mike has let us know
that this should happen 10 days after we take our cut of data
(on the following Thursday).

This commit makes sure we use the second most recent set of statistics
until the publish date for the current month.