Move logic for rendering the correct set of exports out of the controller
This adds in a #latest_publishable_exports method to the
MonthlyStatisticsReport model which returns the latest export for each
of the MONTHLY_STATISTICS_EXPORTS.

It then uses locales to build the link text.

It then ensures that the latest exports are returned, with the exception
of the period of time between generation and publish for the current
month. During this period it will continue to return the previous months.