Move the complaint link up, away from the other links
In research, the "Make a complaint, give feedback, or let us know if you’ve been treated unfairly" in the footer was often not seen.

By moving it to before the other links, and separating it with some white space, we hope to make it more visible.

The link also now semantically is contained within the "Get help" headed-section rather the "Support links" invisibly-headed section, which should improve accessibility.