Stub ProcessState#state result
This isn't the focal point of the spec in question, we have had sporadic failures on this test which appear to be to do with comparing the created_at vs. updated_at timestamps on the application form.
It's likely that a comparison of DateTime.to_i can sometimes fail because the two values fall either side of a second time lapse.
In any case it's not relevant to this test which asserts the order of results.