Fix ApplicationsByDemographicDomicileAndDegreeClassExport columns
Following feedback from TAD on this export it was deduced that the
original requirements did not include the adjusted columns. This change
removes these. Also, when running the report on Prod I noticed that null
values for ethnicity are being pulled through so I've added a line to the
SQL query to filter these out