Stop multiple PR build triggers
PR builds are being triggered multiple times. This is because you get
one for the creation of a PR then one per label.. this is mainly
problematic for dependabot PRs.

This commit attempts to fix that by adding some new conditions at the
job level to ensure that only certain conditions trigger builds.