Remove logic to add a year when calculating age range in spec
We calculate the age according to the beginning of August and end
of July when generating reports. However if we change the logic in
the specs we risk having to generate different results each time.

Remove the logic around adding a year after August to keep the
spec expectation the same.