Use MonthlyReport test helper to test Ministerial report
The report specs now use the standard monthly statistics dataset too, which
is good because we can see how inconsistencies propagate to each of our
different reports if we change that core data set.

We add multiple application choices to MonthlyStatisticsTestHelper data,
so that we can test that part of the Ministerial report too (and
guarantee no regressions in the external report).

We also ensure that MonthlyStatisticsTestHelper always provides a
subject for courses. This is necessary because we join on this in the
Ministerial Report, so without it applications to courses without
subjects will be filtered out.

This done, fix the MinisterialReport specs so they expect the right
numbers of applications now that we're using the helper to generate test
apps.

Now that we are collecting FE applicants, make sure these are
consistently bucketed into "Secondary" when we determine subjects.

(Not technically true, but apparently this is how we do it.)