Add a show action
Extract show duplicate match table to a component
We decide to create a small struct for a Duplicate Match so
we can show all attributes in a clean way in the view.
Another decision was to loop through the table rows over and over
for explicit code instead of a more helper dynamic method.