Rename Yabeda exported_instance label
We currently send the cloud foundry instance id as "exported_instance" label
so it doesn't override the "instance" label generated by prometheus. But we
want to drop the prometheus one because it changes at each deployment
and creates millions of time series in prometheus and influxdb.

Instead we're going to send the instance id as "app_instance" which is
used to replace the prometheus instance label as configured in
cf-monitoring.