Guard against nil last names on last application forms
It has been reported that when prefilling applications on sandbox, applications
might have a nil last name sometimes. This is because we are filling in the last
name if the duplicate trait is sampled with the last name from the last application.
However an application have have this still not filled in, meaning we will be leaving
the last name blank

Make sure to use the last name only if the last application form's last name is
present.