Fix evaluation of subjects per candidate
Was picking subjects from the first (apply 1) application. This changes
the strategy to pick from the latest application form (be it the
original or a subsequent apply 2 one). Doing this makes the evaluation
of subject consistent with that of state (status).

Also 'Fix' the spec for ministerial report per subject/status. This just
corrects the spec to fit the current implementation and (more
importantly) the latest test data. Just one test changed at this point
as a starting point to be able to build tests for further changes.