Add pagination to duplicate matches support interface
The previous and next link was customized removing the arrows and
the only graceful way was to override the pagy_t method in
the duplicate matches table component.

Also another decision that I made is that if someone wants to type
in the url and inexisting page (e.g ?page=100000) the app will redirect
to the last page.