Sanitize `last_saved_state` to avoid deep merging further conditions
Every time we update further conditions and attempt to deep merge
it into the previous state, we are maintaining old values which is
causing a mess in the final result. This only happens when JS
is enabled as we leave the recalculation of the conditions to
when its submitted, as opposed to disabled javascript which
deals with those individually on each request.

To resolve the issue we are resetting the last saved state
further conditions if there are any updates, to only rely on
attrs conditions as the source of truth.

To ensure we don't set unpermitted params always change the
attrs to a hash before attempting to read the hash values.

We also check for the key of the further conditions in the attrs
rather than presence to ensure that when we remove all conditions
we also reset the last saved state.