Use current_course_option for course level in TAD export
Previously we were extracting `course.level` where `course` was
delegated to `application_choice`. We should be using
`current_course_option` or `current_course` for all course data that we
export for this report so that we reflect the course that the candidate
ended up being offered over then one that they originally applied for
(on the odd occasions that they are different).