Change candidate account status on support interface
With this change the support user can block/unblock a submission,
unlock/lock an account or unblocked the candidate account.

Caveats:
We have problems on i18t with the gov uk radios to look automatically
to config/locales/en.yml because we are using the same attribute
(status).
So we had to use the human_attribute_name from Active Model model
directly.