Make order of subjects explicit in SQL reports
Adding a method to sort subjects by `id` before determinining the
dominant subject only works for reports that deal in Activerecord
objects. There are a couple of reports that use raw SQL and aggregate
the subject names and codes themselves. This commit just adds an
explicit ordering to those aggregates so that they also have a
deterministic behaviour wrt subject ordering.