Mark new matches & new duplicate candidates as submission blocked
We should not mark candidates from existing matches as submission
blocked.

Let's say one day we create a match with 2 candidates and both are
mark as submission blocked.

The support team unblocks one account and leave the other blocked.

If the job runs again we should not mark everything blocked again.

We should only mark as submission blocked, new matches and new
candidate duplicates.

This commit address this issue