Mark duplicate candidates as submission blocked and send emails
We need to send the duplicate match emails to the duplicate matches
candidates from the period of January 26th - February 2nd which is the
period that we turn off the duplicate matching feature flag and
those new matches weren't notified.

Currently, there are 21 matches in this period.
The candidates on these matches should receive the duplicate match email
and being marked as submission blocked.

We need to mark the duplicate matches
candidates as submission blocked from the period of
January 25th - February 2nd which is the
period that we turn off the duplicate matching feature flag and
those candidates weren't marked as submission blocked