Properly handle multiple apply again applications
We are seeing discrepancies in ministerial reports when there are
multiple apply again applications the last of which is unsubmitted. It
looks like the code assumes that all apply again applications are
unsubmitted though this isn't always the case. We should discount any
unsubmitted applications.

We should also not assume that apply again applications can only have a
single application choice.