Use `find_each` to reduce memory consumption
The query for TAD data export is timing out. I don't think this is the
reason but it makes sense to avoid loading half the database into memory
at once.

Also changes the `preloads` to match the associations that are actually
needed by this report (this bit needs testing properly).