WIP
Attempt of making the rows of each record to be like a Hash and also
like a Struct.

I could use Struct but then we need to map the schema without incurs
in exceptions.

For example if we use the first row as base for schema:

[{ a: :b}, {a: :b1, c: :d}]

This will incurs in error in Struct as we need to get all properties
from the whole data (this could be slow if has thousands of records)

So hence I used the method missing