Fix rubocop ofence for open struct usage
I am being asked to support dynamic structures that work as a structure
and as a Hash. For simplicity I consciously decided to use OpenStruct
even with so many people disapproving the usage of OpenStruct.

The code to use Struct for dynamic lists that we don't have control
and are dynamic and the expectation that should also behave like a Hash
made the old code too problematic to keep and maintain. And the
OpenStruct was much simple and faster.

I am happy to receive suggestion & alternatives for the problem.