Migration to drop application_forms.address_type not null and default
We want to be able to prompt candidates to specify where they live (in
the UK or elsewhere) without pre-filling either of the options (radio
buttons). However we do need to show the previously selected value if
the candidate visits this form a second or subsequent time.

The simplest way to achieve this is to remove the not null and default
constraints on the database column.