Added step to evaluate job status
The workflow reports a successful deployment if the job.status is
evaluated as successful. The workflow includes steps that run even if
preceeding steps fail, this means the job.status can be successful even
when some steps have failed. This change adds a step to evaluate the
outcome of each step and report a failed deployment if one of them has
failed.