Create the reference data library
This is an extract from Alaric's repository and moved into this
app temporarily. After the works is done this will be extracted
to your own gem.

------------------------------------
The plan with the gem reference data
------------------------------------

1. Once the gem is in reasonable shape (which should be soon),
copy the whole thing into app/lib in Apply.
2. Try to integrate with apply, amending the gem along the way>
3. Once stable, extract gem and add as dependency.

--------------------------------------------------
In regards to the DfE::ReferenceData::Record class
--------------------------------------------------

I am being asked to support dynamic structures that work as a structure
and as a Hash. For simplicity I consciously decided to use OpenStruct
even with so many people disapproving the usage of OpenStruct.

The code to use Struct for dynamic lists that we don't have control
and are dynamic and the expectation that should also behave like a Hash
made the old code too problematic to keep and maintain. And the
OpenStruct was much simple and faster.

----------------------
In regards to the data
----------------------

- `QUALIFICATIONS` includes `level` as a field, so that information is still available
 - `TYPES` refers to qualification IDs that can be looked up in `QUALIFICATIONS`
 - Qualifications and their levels taken from
   https://www.gov.uk/what-different-qualification-levels-mean/list-of-qualification-levels

Co-authored-by: Alaric Snell-Pym <alaric@snell-pym.org.uk>