Show old duplicate matches from all recruitment cycles
This can help the support team to solve accounts that were created
last recruitment cycle and it was submitted this recruitment cycle.

Example:

Two accounts were created in 2021 on the last recruitment cycle,
created as a match in the last recruitment cycle, but it passes some
months and both of the accounts submitted an application this year.

This commit solves that.

Co-authored-by: Steve Hook <steve.hook@gmail.com>