Revert the renaming of `AddIndexToProviderUserDfESignInUid`
This migration class name was changed to `AddIndexToProviderUserDfeSignInUid` to appease Rubocop. This seems to cause errors when deploying review apps, e.g.:

https://sentry.io/organizations/dfe-teacher-services/issues/3110289345/?project=1765973&referrer=slack

I think the ActiveSupport Inflector is expecting the file name to map to
the old class name because `DfE` is declared as an acronym in
`config/inflector.rb`.