Make reference status and reminder link more prominent
We have some evidence that candidates are adding multiple referees with the same email address, often a few days apart.

We think this may be because they are doing this in order to re-request a reference from a referee who has not yet responded, and that they have not realised that there is a feature to send a reminder.

This aims to reduce the number of error messages about adding a referee with the same email as an existing one by moving the status row to the top of the summary, and moving the "Send a reminder to this referee" link to the top, next to the status.