Do not let enum override ApplicationChoice#rejection_reason?
This caused a bug where calling code expected this to be false, because
there was no value in that field. In fact Rails enum magic created a
method from the value of the rejection_reasons_type enum with the same
name - and this returned true because there was in fact a value in the
rejection_reasons_type field that we had forgotten to revert.

Fix the spec and cause the revert to actually nil out this field.