Attempt to fix flakey app choice filtering spec
I've seen spec failures a couple of times and they seem to be due to
the randomness of values assigned in the application form factories.

If we create two application forms via factories there is a slim chance
that both will be assigned the same last name. This breaks at least one
of the examples in this test file.

Faker has a `unique` option but I don't think we can use that in
factories because it would limit the number of objects we can create
without errors (Faker throws an error if it runs out of unique values).