Handle missing end dates gracefully in JobComponent
This component was updated to use `Date#to_fs` rather than `to_s` as
part of the Rails 7 upgrade. This causes issues when `end_date` is `nil`
because `to_fs`, unlike `to_s`, does not handle nil values.

In production there are a number of older work experience records that
have both `end_date` and `currently_working` set to `nil`. I am
proposing here that we handle them in the same way that volunteering
experiences are handled, e.g.
https://github.com/DFE-Digital/apply-for-teacher-training/blob/main/app/components/shared/work_history_item_component.rb#L560

All `start_date` values in the production DB are set so I don't think we
have to do anything similar there.