Search base coverage artifact by name and status.
This is a less performant way of finding the base coverage artifact but
it reduces the impact of issues with GitHub API searches failing to find
by branch name.

The download-artifact workflow action iterates over all workflow runs
until it finds a successful run with the artifact with the correct name.