Allow multiple application choices at different sites
No longer raise an alert for applications with multiple choices of the
same course. We allow this behaviour as long as the site and/or study
mode are different, and there is already model validation in place to
enforce CourseOption uniqueness.

Running a query on production to find applications with duplicate course
options yields no results, therefore decided no value testing for it.