Reimplement return to review page after editing behaviour
This was broken by the previous commit. It's no longer sufficient to
check whether a record exists to determine whether it's a new/edit
operation. Instead we use the existing `from` param to work out whether
we need to return to the review page or the next 'page' in the form
sequence.