Increase char limit for job role from 60 to 100
Analysis of error messages shows that candidates often slightly exceed the (unstated) character limit by a small amount, for reasonable-sounding job titles.

They also sometimes vastly exceed the limit by adding a whole paragraph describing the job (which the field is not intended for). A separate Pull Request will alter the label to try and avoid this.