Make error type visible to the layout
The support layout uses the page title as a H1 heading but error pages
already contain this heading in the template body.
We prevented a double rendering of the H1 title by checking if the
controller_name was 'errors' but this does not work for
ActiveRecord::QueryCanceled errors emanating from the
ApplicationController as we explicitly want the actual controller name
to propagate into error tracing.

So be explicit about the error type and check for the presence of this
in the support layout.