Remove terraform variable key_vault_cypress_secret_name
The value is not used by terraform and causes a warning. It's not really
required anyway as the value is the same for all environments. We
simplify by hardcoding it instead.

Terraform warning:

```
Warning: Value for undeclared variable

The root module does not declare a variable named
"key_vault_cypress_secret_name" but a value was found in file
"workspace_variables/production.tfvars.json". To use this value, add a
"variable" block to the configuration.

Using a variables file to set an undeclared variable is deprecated and will
become an error in a future release. If you wish to provide certain "global"
settings to all configurations in your organization, use TF_VAR_...
environment variables to set these instead.
```