Scope site lookup for vendor API to current cycle
A provider through the API can supply course and
site attirbutes to help us map it to the correct
course option to make or change an offer to.

If we `find_by` sites across all cycles this will yield
incorrect error messages. Scope the lookup to current cycle
year to ensure we are only looking at the valid sites.