Mark record invalid if there are duplicate codes for site
We no longer rely on code being unique per provider per cycle.
Add an error message if we encounter two sites using the same
code for a provider in the same cycle.

This is an edge case, but this will help debugging the issue,
rather than raising the other error around the course option
not found.