Update to govuk-components 3.0.6
Due to some breaking changes, this requires also updating the dfe-autocomplete gem, and updating some of the code.

The changes are:

* setting `classes` to be an empty array instead of `nil` for some component items
* setting the :id attribute within an `:html_attributes` param for the govuk_accordion component
* updating the tests for secondary buttons so that they are less brittle and don’t fail when additional HTML attributes are present

These tests were failing due to an update to the `govuk-components` gem which added these missing attributes to links styled as a a button: `role="button" draggable="false" data-module="govuk-button"`.

(These are required for accessibility, to allow the links to behave more like buttons so that "click button" commands still work).

I've updated the tests to use the `have_link` matcher rather than expecting a raw string, as this means they'll still work regardless of any extra attributes and attribute order.