Add worker to update out of date provider ids on applications
If an accredited provider changes in publish, existing applications
to courses will hold stale provider id data. Add a worker which finds
out of date applications and updates the provider ids appropriately.

We do not want to trigger any callbacks on application choices not from
this cycle, add an extra check and use update_columns to avoid callbacks
which will amend the `updated_at` timestamp on an application choice.