Add step to automerge dependabot prs
If a dependency upgrade is done by dependabot, and passes
all test, linting and deployment to review app steps merge
it automatically to main.

We are skipping govuk frontend dependencies to ensure they are done
manually as sometimes they have recommendations we should address
before merging, although they might not fail the build.

To make sure we trigger workflows, use the PAT token with the required
permissions instead of GITHUB_TOKEN