Mark incomplete references as 'selected'
These applications are to be used by vendors to check their integrations work with
the new reference flow, thus we need to mark references as 'selected' so
that they appear in the `GET applications` and `GET applications/:id`
responses.

Also add a parameter to the `GET applications` and `GET applications/:id`
responses so that they can optionally include references for which
feedback has not been provided.