Honour incomplete references when caching applications
The Vendor API caches the serialized json as well as the presented hash for each application, but these can differ depending on the incomplete_references param in the request which causes applications with incomplete references to appear when they were not expected and vice versa. This commit reworks the cache key to include a hashed hash of key suffixes, ie. a predictable integer generated from the contents of a hash.
eg. { method: :as_json, incomplete_references: false }.hash -> 3410546498469832963
This way we end up with predictable cache key suffix depending on the method (as_json or not) and the incomplete references param value.