Improved content for start of cycle emails
This is part of tidying up the end of cycle content. I've made the language more consistent with other parts of the service.

Also we now link on text rather than showing URLs, as the URLs have tracking information and are much longer. I suggest linking on the action which the user is taking.