Update privacy policy for new references process
We no longer use the term 'referee' in the service so I've changed it to 'person giving a reference'.

I've also made changes to correctly reflect the data we gather and show. In particular, we will show the name of the training provider to the person giving a reference. 

I've tried to keep changes to a minimum as there are a lot of ways in which this content could be improved.