Deleted 'where you work'
Candidates do not necessarily tell us where someone works when they give us reference details. They might do as part of the 'how you know them' question, but  it may not be relevant if it's a character reference. So I've deleted it since we already have 'how you know them' as a separate bullet point.