Updated provider version of privacy policy
The Manage service has its own version of the privacy policy, using markdown (whereas the candidate version uses html). The content should be the same.

I have not compared the two versions as that isn't part of the current work. I've just made the same changes in the Manage version as I made in the Apply version (which is in the same pull request).