Fix a flakey spec in `DetectInvariantsDailyCheck`
When selecting records from an SQL database there's no guarantee of
returned order if one is not specified.

If we were just checking two arrays in this test then we could use a
matcher that ignores order, but since it's converted to a sentence that's
not possible.

Ordering by name seems as good an option as any.