Simplify the candidate sign-up smoke test
We don't really need to be checking Notify's API to get the body of the
signup email; we can use local test delivery and extract the link that
way.

If we really need a smoke test to confirm our integration with Notify
then we can add one explicitly for that, and use Notify's proper smoke
test procedures.