Don't generate interviews for test applications when between cycles
While the `TestApplications` class manages its own copy of a running
timestamp, other parts of the application still use `Time.zone.now`,
especially the `CycleTimetable`, so various parts of the test application
creation is still tied to the real world time.  In this particular case,
the scheduling of interviews.

A thorough revision of how we try to artificially manage time outside of
the test suite is needed to fix this properly, for now we're just
skipping the creation of interviews if we're between cycles.

This will start working normally again after Find reopens.

Co-authored-by: Tomas D'Stefano <tomas_stefano@successoft.com>