Make rack-mini-profiler opt-in
This gem slows down each request when providing profiling information,
so it makes sense to only invoke it when we're going to make use of that
information.

Set the env var RACK_MINI_PROFILER to `true` when you want to see
profiling information (or if you want to put it back to how it was).