Remove content from reference request page
This removes 2 lines of content from the reference request page shown only when requesting additional references after an offer has been accepted.

We’ve decided to remove these lines as:

* we no longer use the term "referee" anywhere in the service
* we don’t think we need to explain that referees cannot see the content of applications, as this hasn't come up in research as a concern. This is also still explained in the privacy policy.
* we don’t think candidates need telling to contact their referees, as many will anyway, and if they do, they don’t necessarily need to do it before sending the request, but could do it straight away after

If the need for either of these pieces of content re-emerges, we can re-review, also adding this content in the other places where references are requested.