Updated conditions_met email
This email said that the provider had "confirmed that... they've checked your references". This is incorrect.

I've used the latest wording for this email, including a link so the candidate can check on their references. We don't know whether or not the provider is happy with the references.

I've also added the content on deferring from the unconditional_offer email. It also applies here.