Updated reinstated_offer
I used the most recent version of the content. In particular, I don't think we should mention references in this email because:

- in order for the offer to have been deferred, the candidate must have requested references in the previous cycle - in this first cycle of the new process, we know that they must also have received their references (using the old process)
- If the candidate signs in, they will not see any information about references

As a result, it's better to leave references as an off-service process in this situation. If we later add a way to mark references as complete then we may want to think more about this.

I've removed the logic based on whether this is using the new process. There's no point giving a link to the service to users with a reinstated offer who have no conditions, as there's nothing important for them to see when they sign in. Once I put the link into the logic for whether there are conditions, there was no meaningful distinction between the old and new process emails.