Updated unconditional_offer_accepted
I've used the most recent version of the content for this email. 

In particular, the new content mentions references. This is because the requests will only be sent when the unconditional offer is accepted, so we know that the provider will not have received them when this email is sent.