Update wording of ethnicity questions
This updates the wording of the Ethnicity questions to match that in the [Equality information pattern](https://design-system.service.gov.uk/patterns/equality-information/) in the GOV.UK Design System, which is in turn based on the [Ethnicity harmonised standard](https://analysisfunction.civilservice.gov.uk/policy-store/ethnicity-harmonised-standard/) from the Government Analysis Function.

The changes are mostly ordering changes, ie the ordering of the groups in the options lists, and the ordering within the wording (eg "Irish Traveller or Gypsy" becomes "Gypsy or Irish Traveller").

"Another" has also changed to "Any other" and "Another Mixed background" has changed to "Any other mixed or multiple ethnic background".