Don't create duplicate candidates in test applications
Currently 1 in 4 test applications created within review apps are deliberately created as "duplicate candidates" which are then blocked from being able to apply.

Most of the time, we aren’t testing this feature, and it can get in the way of testing the apply again feature (you have to manually unblock the candidate first).