Fix a flakey spec (again)
This is a weird one.  In theory it should be consistent because we're
using `.order(:name)` in the SQL, but turns out postgres delegates its
string sorting to the local C library, so I guess it's possible different
workers are using different libraries?  Different locales?  Different
collations?  I dunno but it's not worth digging into further unless we
have a feature that really definitely needs a specific string sort order.

See the (somewhat jaded) FAQ entry:
https://wiki.postgresql.org/wiki/FAQ#Why_do_my_strings_sort_incorrectly.3F