Don't try to render grade details for qualifications marked as missing
Qualifications can be of a type "missing" where the candidate for one
reason or another did not get a required qualification (e.g. a core GCSE).

In those cases we don't have any grades to show, so we should not
attempt to render any.

This is a relatively rare issue as it only caused a problem rendering
the science tripe award, which is why it only showed up now.