Remove the references email address index
On the error message for email address on references there is one request from product to allow duplication if the other reference
is cancelled (example: if you have a cancelled
reference “jollygoodfellow@gmail.com” and you try to add another
reference on post dashboard with the same email we should
have two records).

This is fine although we have an unique index on references table for the reference_id and email address so it blows up:
CandidateInterface::Reference::RequestRefereeEmailAddressForm validations when a duplicate email is given when duplicate reference is cancelled returns custom error message

     ActiveRecord::RecordNotUnique:
       PG::UniqueViolation: ERROR:  duplicate key value violates unique constraint "index_references_on_application_form_id_and_email_address"
       DETAIL:  Key (application_form_id, email_address)=(22188, iamtheone@whoknocks.com) already exists.

So in order to accomodate the change to accept duplicate emails in
certain conditions we need to remove the index.