Sandbox only: update course UUIDs
We have seen an issue on Sandbox whereby courses aren't showing up,
because `exposed_on_find == false`, which doesn't correspond to the
`exposed` property on the Publish API. It turns out the sync job is
failing due to a validation error.

When the sync job syncs a given course from the API, it tries to find a
match on the Apply database by uuid and recruitment_cycle_year. However, as
Sandbox Publish courses have changed UUIDs following a
backup-and-restore from Production, course UUIDs are no longer matching.
Hence we get a validation error when trying to create a new course with
the same provider, course code and recruitment cycle.

This data migration runs on Sandbox only to update Course#uuid to match
the corresponding value on Publish. Once it has run, a full Publish API
sync should update the rest of the data.